requirements.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.7.2
astunparse==1.6.3
billiard==4.1.0
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-jazzmin==2.6.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
keras==2.13.1rc0
kombu==5.3.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
